refactor: 그룹 생성 서비스 코드 리팩토링
