--- domain/generator/src/main/kotlin/com/few/generator/domain/vo/DateTimeRange.kt
@@ -1,8 +1,0 @@
-package com.few.generator.domain.vo
-
-import java.time.LocalDateTime
-
-data class DateTimeRange(
-    val startTime: LocalDateTime,
-    val endTime: LocalDateTime,
-)
